# ENH: Don't cythonize pure-python GBT conversion code