--- a/generator/gen_daal4py.py
+++ b/generator/gen_daal4py.py
@@ -1255,8 +1255,6 @@ def gen_daal4py(dalroot, outdir, version, warn_all=False, no_dist=False, no_stre
     ):
         with open(jp("src", "gbt_model_builder.pyx"), "r") as f:
             pyx_gbt_model_builder = f.read()
-        with open(jp("src", "gbt_convertors.pyx"), "r") as f:
-            pyx_gbt_generators = f.read()
     if (
         "algorithms::logistic_regression" in iface.namespace_dict
         and "ModelBuilder"