## Description

Currently, there is a .pyx file under `src/` containing the code for the model builders from xgb/lgb/cb. This code is written in pure python without using any cython types or functions, hence it doesn't see any speed ups from being compiled through cython. Moreover, since the code deals with python objects like strings, passing it through cython might actually cause slight slowdowns.

This PR moves the code to pure-python .py files. Benefits:
* Reduced compilation times.
* Better error traces.
* Easier to modify.

As a comparison, I timed a model conversion from XGB from the model fitted on the bosch dataset from the regular config in the benchmarks using the jupyter 'timeit' tool.
* Before (current main): 47.7 ms ± 2 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
* After (this PR): 45.2 ms ± 1.2 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

So as can be seen, there is no performance degradation from this move.

---

PR should start as a draft, then move to ready for review state after CI is passed and all applicable checkboxes are closed.
This approach ensures that reviewers don't spend extra time asking for regular requirements.

You can remove a checkbox as not applicable only if it doesn't relate to this PR in any way.
For example, PR with docs update doesn't require checkboxes for performance while PR with any change in actual code should have checkboxes and justify how this code change is expected to affect performance (or justification should be self-evident).

Checklist to comply with **before moving PR from draft**:

**PR completeness and readability**

- [x] I have reviewed my changes thoroughly before submitting this pull request.
- [x] Git commit message contains an appropriate signed-off-by string _(see [CONTRIBUTING.md](https://github.com/uxlfoundation/scikit-learn-intelex/blob/main/CONTRIBUTING.md#pull-requests) for details)_.
- [x] I have added a respective label(s) to PR if I have a permission for that.
- [x] I have resolved any merge conflicts that might occur with the base branch.

**Testing**

- [x] I have run it locally and tested the changes extensively.
- [x] All CI jobs are green or I have provided justification why they aren't.

**Performance**

- [x] I have measured performance for affected algorithms using [scikit-learn_bench](https://github.com/IntelPython/scikit-learn_bench) and provided at least summary table with measured data, if performance change is expected.