@david-cortes-intel 2025-03-31
I doubt the CI failure is related to the changes here:
````
  File "/usr/share/miniconda/envs/build-env/lib/python3.11/site-packages/conda_build/os_utils/liefldd.py", line 381, in <listcomp>
    result = [e.name for e in dynamic_entries if e.tag == ELF_DYNAMIC_TAGS.SONAME]
                                                 ^^^^^
ValueError: 32212254717 is not a valid TAG.
```